+++ worker/funfact.go
@@ -31,7 +31,7 @@ func (d *Dependency) FunFact(ctx context.Context, in *pb.Member) (*pb.FunFactRes
 
 	// Read about buffered channel vs non-buffered channels
 	wpm := make(chan uint32, 1)
-	deletionRate := make(chan float32, 1)
+	deletionRate := make(chan float64, 1)
 	attempt := make(chan uint32, 1)
 
 	// Run all the calculate function concurently
@@ -63,7 +63,7 @@ func (d *Dependency) FunFact(ctx context.Context, in *pb.Member) (*pb.FunFactRes
 
 	var result = struct {
 		Wpm          uint32  `json:"wpm"`
-		DeletionRate float32 `json:"deletion_rate"`
+		DeletionRate float64 `json:"deletion_rate"`
 		Attempt      uint32  `json:"attempt"`
 	}{
 		<-wpm,
